src/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
mysql-connector-python==8.0.29
google-auth==2.9.0
requests==2.28.1